+++ CHANGELOG.md
@@ -11,8 +11,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Fixes
 
 - Improved pool name
+- Improved partial swap result
 
 ### Perfomance
+
 
 ## 0.1.2 - 2021-09-29